[BE] feat: 리뷰 모아보기 API 구현 (#806)
* refactor: dto 이름 변경

AnswerContentResponse -> TextResponse

* feat: 질문ID에 해당하는 Answer 반환 함수 추가

* feat: 질문ID에 해당하는 OptionItem들 반환 함수 추가

* feat: 리뷰 요청 코드와 섹션ID에 해당하는 질문 반환 함수 추가

* feat: 리뷰 모이보기 API 구현

* refactor: 테스트에서 검증하고자 하는 것을 분명히

* refactor: 패키지 의존 방향 통일을 위한 레포지토리 함수 이동

* refactor: 테스트 목적에 해당하는 것만 남기기

* refactor: 함수 이름 변경

- 인자가 List<Long> 이므로, ids 로 끝나게 변경함

* refactor: 가동성을 위한 함수 분리

* refactor: 다른 서비스 클래스들과 네이밍 통일

* feat: 리뷰 요청 코드 검증, 섹션 아이디 검증 추가

* refactor: 섹션에 해당하는 질문 가져오는 함수 수정

AS-IS: 내 리뷰그룹 중 특정 세션에 대한 질문 가져오기
TO-BE: 리뷰그룹 검증 완료되었으므로 특정 세션에 대해서만 가져오도록 수정

* refactor: 답변 가져오는 함수 수정

AS-IS: 특정 질문에 대해서 답변된 것 다 받아오기
TO-BE: 내 리뷰 그룹에 해당한다는 조건 추가

* feat: 답변하지 않은 내용은 빈 배열로 받아오도록 하는 기능 추가

* fix: 깨지는 테스트 코드 봉합

* refactor: Mapper 분리

* refactor: 섹션 검증 방법 변경

AS-IS: boolean 으로 검증, sectionId 인자 그대로 사용
TO-BE: Optional<Section>으로 검증, section.getId() 사용

* refactor: 필요없는 JOIN 제거

* test: 테스트 코드 목적별로 분리

* feat: 질문 목록을 position 순서대로 정렬

* refactor: 테스트 코드 extracting 사용 변경

AS-IS: attribute 값을 문자열로 하드코딩
TO-BE: dtp 의 필드명으로 가져오기 e.g. VoteResponse::content

* chore: 오타 수정

* test: 다른 리뷰 그룹이 있는 상황에 대한 테스트

* refactor: 접근제어자 수정

* feat: 하이라이트 응답 추가

* test: 추가된 속성 반환 테스트

* refactor: 유연한 레포지토리 함수로 변경

* refactor: 변수명 변경

* refactor: 가독성 개선

함수 분리, 변수명 변경, 함수 인자 변경

* refactor: Answer -> 구체Answer 캐스팅 예외 추가

* refactor: 서술형 답변 가져오는 함수 수정

* refactor: 선택지 목록 가져오는 함수 수정

* test: reviewGatherMapper에 대한 테스트 작성

* style: 개행

- 가독성 개선와 컨벤션 유지를 위함

* chore: 로그 메세지 수정

- 컨벤션 유지를 위함

* refactor: 컨트롤러 인자 타입 변경

- Long -> long

* feat: 최대로 내려주는 응답 수 제한 기능 구현

* refactor: 테스트 코드 필드 접근제한자 추가

* refactor: 변수명 변경

* chore: 사용하지 않는 메서드 제거

+++ backend/src/main/java/reviewme/review/service/dto/response/gathered/VoteResponse.java
@@ -2,6 +2,6 @@ package reviewme.review.service.dto.response.gathered;
 
 public record VoteResponse(
         String content,
-        int count
+        long count
 ) {
 }
